requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-images
scikit-learn
tensorflow>=1.3.0
keras>=2.1.6
opencv-python
h5py
imgaug
flask
IPython[all]